feat: add new gas costs (#3813)

+++ internal/pkg/types/message.go
@@ -150,12 +150,12 @@ func (msg *UnsignedMessage) Cid() (cid.Cid, error) {
 }
 
 // OnChainLen returns the amount of bytes used to represent the message on chain.
-func (msg *UnsignedMessage) OnChainLen() uint32 {
+func (msg *UnsignedMessage) OnChainLen() int {
 	bits, err := encoding.Encode(msg)
 	if err != nil {
 		panic(err)
 	}
-	return uint32(len(bits))
+	return len(bits)
 }
 
 func (msg *UnsignedMessage) String() string {
